refactor: member/에서 auth/로 인증 로직 이동 및 클래스명 변경
