+++ src/main/kotlin/com/wq/auth/api/controller/auth/request/EmailLoginRequestDto.kt
@@ -1,3 +1,3 @@
-package com.wq.auth.api.controller.member.request
+package com.wq.auth.api.controller.auth.request
 
 data class EmailLoginRequestDto (val email: String, val verifyCode: String, val deviceId: String?)
