+++ src/main/kotlin/com/wq/auth/api/domain/auth/AuthProviderRepository.kt
@@ -1,8 +1,8 @@
-package com.wq.auth.api.domain.member
+package com.wq.auth.api.domain.auth
 
-import com.wq.auth.api.domain.member.entity.AuthProviderEntity
+import com.wq.auth.api.domain.auth.entity.AuthProviderEntity
 import com.wq.auth.api.domain.member.entity.MemberEntity
-import com.wq.auth.api.domain.member.entity.ProviderType
+import com.wq.auth.api.domain.auth.entity.ProviderType
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.Optional
