+++ src/main/kotlin/com/wq/auth/api/domain/auth/RefreshTokenRepository.kt
@@ -1,7 +1,7 @@
-package com.wq.auth.api.domain.member
+package com.wq.auth.api.domain.auth
 
 import com.wq.auth.api.domain.member.entity.MemberEntity
-import com.wq.auth.api.domain.member.entity.RefreshTokenEntity
+import com.wq.auth.api.domain.auth.entity.RefreshTokenEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
