+++ src/main/kotlin/com/wq/auth/api/domain/member/error/MemberException.kt
@@ -1,5 +1,6 @@
 package com.wq.auth.api.domain.member.error
 
+import com.wq.auth.api.domain.member.error.MemberExceptionCode
 import com.wq.auth.shared.error.ApiException
 
 class MemberException(
