+++ src/main/kotlin/com/wq/auth/domain/auth/request/SocialLoginRequest.kt
@@ -1,6 +1,6 @@
 package com.wq.auth.domain.auth.request
 
-import com.wq.auth.api.domain.member.entity.ProviderType
+import com.wq.auth.api.domain.auth.entity.ProviderType
 
 data class SocialLoginRequest(
     val authCode: String,
